+++ adc_dma_test/adc_dma_test.ioc
@@ -30,10 +30,15 @@ Dma.ADC1.0.Priority=DMA_PRIORITY_LOW
 Dma.ADC1.0.RequestParameters=Instance,Direction,PeriphInc,MemInc,PeriphDataAlignment,MemDataAlignment,Mode,Priority
 Dma.Request0=ADC1
 Dma.RequestsNb=1
+EdgeImpulse.OS-APPSC.1.0.10.IPParameters=MachineLearningCcOSAaAPPSCJjCore
+EdgeImpulse.OS-APPSC.1.0.10.MachineLearningCcOSAaAPPSCJjCore=true
+EdgeImpulse.OS-APPSC.1.0.10.OSAaAPPSCJjMachineLearning_Checked=true
+EdgeImpulse.OS-APPSC.1.0.10_SwParameter=MachineLearningCcOSAaAPPSCJjCore\:true;
+FREERTOS.BinarySemaphores01=dataAvailable,Dynamic,NULL
 FREERTOS.FootprintOK=true
-FREERTOS.IPParameters=Tasks01,FootprintOK,Queues01,configUSE_NEWLIB_REENTRANT,configUSE_TIMERS,Timers01
+FREERTOS.IPParameters=Tasks01,FootprintOK,Queues01,configUSE_NEWLIB_REENTRANT,configUSE_TIMERS,Timers01,BinarySemaphores01
 FREERTOS.Queues01=inferenceTaskQueue,16,uint32_t,0,Dynamic,NULL,NULL
-FREERTOS.Tasks01=samplingTask,2,128,StartSamplingTask,Default,NULL,Dynamic,NULL,NULL;inferencingTask,-3,128,StartInferencingTask,Default,NULL,Dynamic,NULL,NULL
+FREERTOS.Tasks01=samplingTask,2,128,StartSamplingTask,Default,NULL,Dynamic,NULL,NULL;inferencingTask,0,128,StartInferencingTask,Default,NULL,Dynamic,NULL,NULL
 FREERTOS.Timers01=timeoutTimer,timeoutCallback,osTimerOnce,Default,NULL,Dynamic,NULL
 FREERTOS.configUSE_NEWLIB_REENTRANT=1
 FREERTOS.configUSE_TIMERS=1
@@ -61,6 +66,7 @@ Mcu.Pin12=PA11
 Mcu.Pin13=PA12
 Mcu.Pin14=VP_FREERTOS_VS_CMSIS_V1
 Mcu.Pin15=VP_SYS_VS_tim1
+Mcu.Pin16=VP_EdgeImpulse.OS-APPSC_VS_OSAaAPPSCJjMachineLearning_1.0.10_1.0.10
 Mcu.Pin2=PF0-OSC_IN
 Mcu.Pin3=PF1-OSC_OUT
 Mcu.Pin4=PA0
@@ -69,8 +75,9 @@ Mcu.Pin6=PC4
 Mcu.Pin7=PC5
 Mcu.Pin8=PE8
 Mcu.Pin9=PE9
-Mcu.PinsNb=16
-Mcu.ThirdPartyNb=0
+Mcu.PinsNb=17
+Mcu.ThirdParty0=EdgeImpulse.OS-APPSC.1.0.10
+Mcu.ThirdPartyNb=1
 Mcu.UserConstants=
 Mcu.UserName=STM32F303VCTx
 MxCube.Version=6.9.2
@@ -161,7 +168,7 @@ ProjectManager.ToolChainLocation=
 ProjectManager.UAScriptAfterPath=
 ProjectManager.UAScriptBeforePath=
 ProjectManager.UnderRoot=true
-ProjectManager.functionlistsort=1-SystemClock_Config-RCC-false-HAL-false,2-MX_GPIO_Init-GPIO-false-HAL-true,3-MX_DMA_Init-DMA-false-HAL-true,4-MX_ADC1_Init-ADC1-false-HAL-true,5-MX_USB_DEVICE_Init-USB_DEVICE-false-HAL-false
+ProjectManager.functionlistsort=1-SystemClock_Config-RCC-false-HAL-false,2-MX_GPIO_Init-GPIO-false-HAL-true,3-MX_DMA_Init-DMA-false-HAL-true,4-MX_ADC1_Init-ADC1-false-HAL-true,5-MX_USART1_UART_Init-USART1-false-HAL-true,6-MX_USB_PCD_Init-USB-false-HAL-true
 RCC.ADC12PRES=RCC_ADC12PLLCLK_DIV256
 RCC.ADC12outputFreq_Value=281250
 RCC.ADC34outputFreq_Value=72000000
@@ -207,6 +214,8 @@ USART1.BaudRate=115200
 USART1.IPParameters=VirtualMode-Asynchronous,BaudRate,Mode
 USART1.Mode=MODE_TX
 USART1.VirtualMode-Asynchronous=VM_ASYNC
+VP_EdgeImpulse.OS-APPSC_VS_OSAaAPPSCJjMachineLearning_1.0.10_1.0.10.Mode=OSAaAPPSCJjMachineLearning
+VP_EdgeImpulse.OS-APPSC_VS_OSAaAPPSCJjMachineLearning_1.0.10_1.0.10.Signal=EdgeImpulse.OS-APPSC_VS_OSAaAPPSCJjMachineLearning_1.0.10_1.0.10
 VP_FREERTOS_VS_CMSIS_V1.Mode=CMSIS_V1
 VP_FREERTOS_VS_CMSIS_V1.Signal=FREERTOS_VS_CMSIS_V1
 VP_SYS_VS_tim1.Mode=TIM1
